% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenblas.R
\name{last_version_openblas}
\alias{last_version_openblas}
\title{OpenBLAS library versions}
\usage{
last_version_openblas()
}
\description{
OpenBLAS library versions
}
\details{
This function automatically searches \href{https://www.openblas.net/}{\strong{OpenBLAS}} library versions in the official \href{https://github.com/xianyi/OpenBLAS}{\strong{GitHub}} project.
\enumerate{
   \item \code{last_version}: Returns the latest stable version of the \href{https://www.openblas.net/}{\strong{OpenBLAS}} library.
   \item \code{versions}: All stable versions of the \href{https://www.openblas.net/}{\strong{OpenBLAS}} library.
   \item \code{n}: Total number of versions.
}
}
\examples{
last_version_openblas()
}
\seealso{
\code{\link{last_version_r}}, \code{\link{ropenblas}}, \code{\link{rcompiler}}
}
