% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenblas.R
\name{last_version_r}
\alias{last_version_r}
\title{R language versions}
\usage{
last_version_r()
}
\description{
R language versions
}
\examples{
# last_version_r()
}
\seealso{
\code{\link{ropenblas}}, \code{\link{rcompiler}}
}
