% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenblas.R
\name{last_version_r}
\alias{last_version_r}
\title{Given the higher version, the function will return the latest stable version of the \R language.}
\usage{
last_version_r(major = NULL)
}
\arguments{
\item{major}{Major release number of \R language (eg. 1L, 2L, 3L, ...). If \code{major = NULL}, the function
will consider the major release number.}
}
\description{
Given the higher version, the function will return the latest stable version of the \R language.
}
\seealso{
[ropenblas()], [rcompiler()]
\donttest{last_version_r()}
}
