% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenblas.R
\name{ropenblas}
\alias{ropenblas}
\title{Download, compile and configure R to use the OpenBLAS library}
\usage{
ropenblas(x = "0.3.7")
}
\arguments{
\item{x}{OpenBLAS library version to be considered. By default, \code{x = 0.3.7}.}
}
\description{
Link R with an optimized version of the BLAS library (OpenBLAS).
}
\details{
The \code{ropenblas()} function will only work on Linux/Unix systems. When calling the \code{ropenblas()}
function on Windows, no settings will be made. Only a warning message will be issued informing you that the
configuration can only be performed on Unix-like systems.

The function will automatically download the latest version of the OpenBLAS library. However, it is possible to
inform olds versions to the single argument of \code{ropenblas()}. The \code{ropenblas()} function downloads, 
compiles and link R to use of the OpenBLAS library. Everything is done very simply, just loading the library and
invok the function \code{ropenblas()}.
 
Considering using the OpenBLAS library rather than the BLAS may bring extra optimizations for your code and improved
computational performance for your simulations, since OpenBLAS is an optimized implementation of the library BLAS.
}
\note{
You do not have to in every section of R make use of the \code{ropenblas()} function. Once the function is used, R 
will always consider using the OpenBLAS library in future sections.
}
\examples{
# ropenblas()
}
\author{
Pedro Rafael D. Marinho
}
